feat: Criação da tela de cadastro e criando a navegação entre páginas com React Router.
