+++ src/pages/Register.css
@@ -1,0 +1,78 @@
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+    text-decoration: none;
+}
+
+html, body {
+    font-family: "Roboto", sans-serif;
+    background-color: #f5f5f5;
+
+}
+
+.conteiner {
+    background-color: #fff;
+    width: 400px;
+    height: max-content;
+    margin: auto;
+    margin-top: 150px;
+    padding: 20px 15px;
+    border-right: 5px solid rgba(230, 230, 230, 0.504);
+    border-bottom: 5px solid rgba(230, 230, 230, 0.504);
+    border-left: 5px solid rgba(230, 230, 230, 0.504);
+    border-radius: 10px;
+    
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    flex-direction: column;
+}
+
+h1 {
+    margin: 0px;
+    margin-bottom: 40px;
+}
+
+.login-form {
+    padding: 5px;
+}
+
+.form-field {
+    
+    display: block;
+    padding: 5px;
+    width: 200px;
+    margin-top: 20px;
+    margin-bottom: 10px;
+    border: 1px solid #4B4B4B;
+    border-radius: 12px;
+}
+
+.form-error {
+    display: block;
+    color: red;
+    font-size: 0.77em;
+    font-weight: 600;
+    margin-bottom: 10px;
+}
+
+button {
+    
+    margin-top: 20px;
+    margin-bottom: 20px;
+    padding: 5px 10px;
+    width: 200px;
+    border: 1px solid #5C73DB;
+    border-radius: 12px;
+    background-color: #234c97;
+    color: #fff;
+}
+
+button > a {
+    color: #fff;
+}
+
+p {
+    font-size: 0.77em;
+}
